refactor: provide no op relative url resolver by default
