+++ projects/ngx-meta/src/core/src/url-resolution/relative-url-resolver.ts
@@ -1,36 +1,12 @@
-import { InjectionToken, Provider } from '@angular/core'
-import { BaseUrl } from './base-url'
-import { Router } from '@angular/router'
-import { ANGULAR_ROUTER_URL, AngularRouterUrl } from './angular-router-url'
+import { InjectionToken } from '@angular/core'
+import { AngularRouterUrl } from './angular-router-url'
+import { noOpRelativeUrlResolver } from './no-op-relative-url-resolver'
 
 export type RelativeUrlResolver = (url: string | AngularRouterUrl) => string
 
 export const RELATIVE_URL_RESOLVER = new InjectionToken<RelativeUrlResolver>(
   ngDevMode ? 'NgxMeta Relative URL Resolver' : 'NgxMetaRUR',
+  {
+    factory: () => noOpRelativeUrlResolver,
+  },
 )
-
-export const provideRelativeUrlResolver: (baseUrl: BaseUrl) => Provider = (
-  baseUrl,
-) => ({
-  provide: RELATIVE_URL_RESOLVER,
-  useFactory:
-    (router: Router): RelativeUrlResolver =>
-    (givenUrl) => {
-      const relativeUrl =
-        givenUrl === ANGULAR_ROUTER_URL ? router.url : givenUrl
-      /* istanbul ignore next https://github.com/istanbuljs/istanbuljs/issues/719 */
-      if (relativeUrl.length === 0) {
-        return baseUrl
-      }
-      const baseUrlEndsWithSlash = baseUrl.endsWith('/')
-      const relativeUrlStartsWithSlash = relativeUrl.startsWith('/')
-      const pathToAppend =
-        baseUrlEndsWithSlash && relativeUrlStartsWithSlash
-          ? relativeUrl.slice(1)
-          : !baseUrlEndsWithSlash && !relativeUrlStartsWithSlash
-            ? `/${relativeUrl}`
-            : relativeUrl
-      return `${baseUrl}${pathToAppend}`
-    },
-  deps: [Router],
-})
